+++ src/main/scala/pro/civitaspo/digdag/plugin/ecs_task/command/EcsTaskCommandRunner.scala
@@ -15,33 +15,42 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
 
   // For ecs_task.register>  operator (TaskDefinition)
   // NOTE: Use only 1 container
-  // val containerDefinitions: Seq[ContainerDefinition] = params.getList("container_definitions", classOf[Config]).asScala.map(configureContainerDefinition).map(_.get)
-  val additionalContainers: Seq[Config] = params.getListOrEmpty("additional_containers", classOf[Config]).asScala
+  // val containerDefinitions: Seq[ContainerDefinition] = params.parseList("container_definitions", classOf[Config]).asScala.map(configureContainerDefinition).map(_.get)
+  val sidecars: Seq[Config] = params.parseListOrGetEmpty("sidecars", classOf[Config]).asScala
   val cpu: Optional[String] = params.getOptional("cpu", classOf[String])
   val executionRoleArn: Optional[String] = params.getOptional("execution_role_arn", classOf[String])
-  val family: String = params.get("family", classOf[String], params.get("task_name", classOf[String]).replaceAll("\\+", "_"))
+
+  val family: String = params.get(
+    "family",
+    classOf[String],
+    params
+      .get("task_name", classOf[String])
+      .replaceAll("\\+", "_")
+      .replaceAll("\\^", "_")
+      .replaceAll("\\=", "_")
+  )
   val memory: Optional[String] = params.getOptional("memory", classOf[String])
   val networkMode: Optional[String] = params.getOptional("network_mode", classOf[String])
   // NOTE: Use `ecs_task.run>`'s one.
-  // val placementConstraints: Seq[TaskDefinitionPlacementConstraint] = params.getListOrEmpty("placement_constraints", classOf[Config]).asScala.map(configureTaskDefinitionPlacementConstraint).map(_.get)
-  val requiresCompatibilities: Seq[String] = params.getListOrEmpty("requires_compatibilities", classOf[String]).asScala // Valid Values: EC2 | FARGATE
+  // val placementConstraints: Seq[TaskDefinitionPlacementConstraint] = params.parseListOrGetEmpty("placement_constraints", classOf[Config]).asScala.map(configureTaskDefinitionPlacementConstraint).map(_.get)
+  val requiresCompatibilities: Seq[String] = params.parseListOrGetEmpty("requires_compatibilities", classOf[String]).asScala // Valid Values: EC2 | FARGATE
   val taskRoleArn: Optional[String] = params.getOptional("task_role_arn", classOf[String])
-  val volumes: Seq[Config] = params.getListOrEmpty("volumes", classOf[Config]).asScala
+  val volumes: Seq[Config] = params.parseListOrGetEmpty("volumes", classOf[Config]).asScala
 
   // For `ecs_task.register>` operator (ContainerDefinition)
   // NOTE: Set by this plugin
-  // val command: Seq[String] = params.getListOrEmpty("command", classOf[String]).asScala
+  // val command: Seq[String] = params.parseListOrGetEmpty("command", classOf[String]).asScala
   // NOTE: Set in `ecs_task.register>` TaskDefinition Context. If you set it by container level, use the `overrides` option.
   // val cpu: Optional[Int] = params.getOptional("cpu", classOf[Int])
   val disableNetworking: Optional[Boolean] = params.getOptional("disable_networking", classOf[Boolean])
-  val dnsSearchDomains: Seq[String] = params.getListOrEmpty("dns_search_domains", classOf[String]).asScala
-  val dnsServers: Seq[String] = params.getListOrEmpty("dns_servers", classOf[String]).asScala
+  val dnsSearchDomains: Seq[String] = params.parseListOrGetEmpty("dns_search_domains", classOf[String]).asScala
+  val dnsServers: Seq[String] = params.parseListOrGetEmpty("dns_servers", classOf[String]).asScala
   // NOTE: Add some labels by this plugin
   val dockerLabels: Map[String, String] = params.getMapOrEmpty("docker_labels", classOf[String], classOf[String]).asScala.toMap
-  val dockerSecurityOptions: Seq[String] = params.getListOrEmpty("docker_security_options", classOf[String]).asScala
-  val entryPoint: Seq[String] = params.getListOrEmpty("entry_point", classOf[String]).asScala
+  val dockerSecurityOptions: Seq[String] = params.parseListOrGetEmpty("docker_security_options", classOf[String]).asScala
+  val entryPoint: Seq[String] = params.parseListOrGetEmpty("entry_point", classOf[String]).asScala
   // NOTE: Add some envs by this plugin
-  val configEnvironment: Map[String, String] = params.getMapOrEmpty("environment", classOf[String], classOf[String]).asScala.toMap
+  val configEnvironment: Map[String, String] = params.getMapOrEmpty("environments", classOf[String], classOf[String]).asScala.toMap
   // NOTE: This plugin uses only 1 container so `essential` is always true.
   // val essential: Optional[Boolean] = params.getOptional("essential", classOf[Boolean])
   val extraHosts: Map[String, String] = params.getMapOrEmpty("extra_hosts", classOf[String], classOf[String]).asScala.toMap
@@ -49,24 +58,24 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
   val hostname: Optional[String] = params.getOptional("hostname", classOf[String])
   val image: Optional[String] = params.getOptional("image", classOf[String])
   val interactive: Optional[Boolean] = params.getOptional("interactive", classOf[Boolean])
-  val links: Seq[String] = params.getListOrEmpty("links", classOf[String]).asScala
+  val links: Seq[String] = params.parseListOrGetEmpty("links", classOf[String]).asScala
   val linuxParameters: Optional[Config] = params.getOptionalNested("linux_parameters")
   val logConfiguration: Optional[Config] = params.getOptionalNested("log_configuration")
   // NOTE: Set in `ecs_task.register>` TaskDefinition Context. If you set it by container level, use the `overrides` option.
   // val memory: Optional[Int] = params.getOptional("memory", classOf[Int])
   // NOTE: If you set it by container level, use the `overrides` option.
   // val memoryReservation: Optional[Int] = params.getOptional("memory_reservation", classOf[Int])
-  val mountPoints: Seq[Config] = params.getListOrEmpty("mount_points", classOf[Config]).asScala
+  val mountPoints: Seq[Config] = params.parseListOrGetEmpty("mount_points", classOf[Config]).asScala
   val containerName: String = params.get("container_name", classOf[String], family)
-  val portMappings: Seq[Config] = params.getListOrEmpty("port_mappings", classOf[Config]).asScala
+  val portMappings: Seq[Config] = params.parseListOrGetEmpty("port_mappings", classOf[Config]).asScala
   val privileged: Optional[Boolean] = params.getOptional("privileged", classOf[Boolean])
   val pseudoTerminal: Optional[Boolean] = params.getOptional("pseudo_terminal", classOf[Boolean])
   val readonlyRootFilesystem: Optional[Boolean] = params.getOptional("readonly_root_filesystem", classOf[Boolean])
   val repositoryCredentials: Optional[Config] = params.getOptionalNested("repository_credentials")
-  val systemControls: Seq[Config] = params.getListOrEmpty("system_controls", classOf[Config]).asScala
-  val ulimits: Seq[Config] = params.getListOrEmpty("ulimits", classOf[Config]).asScala
+  val systemControls: Seq[Config] = params.parseListOrGetEmpty("system_controls", classOf[Config]).asScala
+  val ulimits: Seq[Config] = params.parseListOrGetEmpty("ulimits", classOf[Config]).asScala
   val user: Optional[String] = params.getOptional("user", classOf[String])
-  val volumesFrom: Seq[Config] = params.getListOrEmpty("volumes_from", classOf[Config]).asScala
+  val volumesFrom: Seq[Config] = params.parseListOrGetEmpty("volumes_from", classOf[Config]).asScala
   val workingDirectory: Optional[String] = params.getOptional("working_directory", classOf[String])
 
   // For ecs_task.run operator
@@ -76,8 +85,8 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
   val launchType: Optional[String] = params.getOptional("launch_type", classOf[String])
   val networkConfiguration: Optional[Config] = params.getOptionalNested("network_configuration")
   val overrides: Optional[Config] = params.getOptionalNested("overrides")
-  val placementConstraints: Seq[Config] = params.getListOrEmpty("placement_constraints", classOf[Config]).asScala
-  val placementStrategy: Seq[Config] = params.getListOrEmpty("placement_strategy", classOf[Config]).asScala
+  val placementConstraints: Seq[Config] = params.parseListOrGetEmpty("placement_constraints", classOf[Config]).asScala
+  val placementStrategy: Seq[Config] = params.parseListOrGetEmpty("placement_strategy", classOf[Config]).asScala
   val platformVersion: Optional[String] = params.getOptional("platform_version", classOf[String])
   val startedBy: Optional[String] = params.getOptional("started_by", classOf[String])
   // NOTE: Generated by ecs_task.register operator
@@ -156,7 +165,7 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
   protected def taskDefinitionConfig(scriptsLocationPrefix: AmazonS3URI): Config = {
     val c: Config = cf.create()
 
-    c.set("container_definitions", (Seq(containerDefinitionConfig(scriptsLocationPrefix)) ++ additionalContainers).asJava)
+    c.set("container_definitions", (Seq(containerDefinitionConfig(scriptsLocationPrefix)) ++ sidecars).asJava)
     c.setOptional("cpu", cpu)
     c.setOptional("execution_role_arn", executionRoleArn)
     c.set("family", family)
@@ -178,7 +187,7 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
     c.setOptional("disable_networking", disableNetworking)
     c.set("dns_search_domains", dnsSearchDomains.asJava)
     c.set("dns_servers", dnsServers.asJava)
-    val additionalLabels: Map[String, String] = Map("pro.civitaspo.digdag.plugin.ecs_task.version" -> "0.0.2")
+    val additionalLabels: Map[String, String] = Map("pro.civitaspo.digdag.plugin.ecs_task.version" -> "0.0.3")
     c.set("docker_labels", (dockerLabels ++ additionalLabels).asJava)
     c.set("entry_point", entryPoint.asJava)
     c.set("environment", (configEnvironment ++ environments).asJava)