+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+0.0.3 (2018-10-30)
+======================
+
+* [Breaking Change] Do not use enum parameter directory because the enums require upper camel case ( `ecs_task.{py,register,run}>` operator)
+* [Enhancement] Rename the configuration key: `additional_containers` to `sidecars` ( `ecs_task.py>` operator)
+* [Breaking Change/Enhancement] Rename the configuration key: `environment` to `environments` ( `ecs_task.{py,register,run}>` operator)
+* [Enhancement] Rename the output key: `last_ecs_task_py` to `last_ecs_task_command` ( `ecs_task.py>` operator)
+* [Fix] Fix example indents
+* [Fix] Avoid java.nio.charset.MalformedInputException: Input length = 1
+* [Fix] Avoid com.amazonaws.services.ecs.model.ClientException: Family contains invalid characters when the default value is used.
+* [Enhancement] Enable to parse json text in configuration
+* [Enhancement] Get s3 content more simply
+* [Fix] Use unique s3 workspace path
+* [Fix] print error in runner.py
+
 0.0.2 (2018-10-29)
 ==================